backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
python-multipart==0.0.9
httpx==0.27.2
sqlalchemy==2.0.34
sqlmodel==0.0.22
passlib[bcrypt]==1.7.4
python-jose==3.3.0
authlib==1.3.1
itsdangerous==2.2.0
orjson==3.10.7
redis==5.0.8
psycopg2-binary==2.9.9
